Сделано создание новой компании с автоматическим созданием ТГ-канала для компании и возможность подключиться к компании через бота

+++ Forms/game_room/create_new.py
@@ -1,19 +1,11 @@
-import asyncio
-import logging
-from aiogram import Bot, Dispatcher, types
+from aiogram import types
 from aiogram.dispatcher import FSMContext
-from datetime import datetime, timedelta
-
-# from aiogram import F
-# from aiogram.fsm.context import FSMContext
-# from aiogram.filters import Command
-# from aiogram.fsm.state import StatesGroup, State
-
-from Schemes.Player import PlayerSheet
 from Tools.MySqlTools import Connection
 from Tools.JsonTools import CatalogJson
 from Tools.BotTools import Tools
-from States import states_connect_to
+from States import states_create_group
+
+import pyro
 
 BotTools = Tools()
 env = CatalogJson(name='file/json/environment.json')
@@ -24,25 +16,36 @@ con = Connection(host=env.read_json_data('DB_host'),
                  password=env.read_json_data('DB_password'))
 
 
-async def linc_group_id(call: types.CallbackQuery):
-    await states_connect_to.StepsConnectTo.id.set()
-    await call.message.answer(f'Введите id компании\n'
-                              f'(попросите вашего Мастера сообщить его вам, если этого не произошло ранее)',
-                              reply_markup=BotTools.construction_inline_keyboard(
-                                 buttons=['На главную'],
-                                 call_back=['start'],
-                                 message=call.message)
-                              )
-    await call.answer()
-    await call.message.delete()
+def get_username(call: types.CallbackQuery):
+    return call.from_user.username
 
 
-async def linc_group_password(message: types.Message, state: FSMContext):
+async def create_group_name(call: types.CallbackQuery):
+    if get_username(call):
+        await states_create_group.StepsCreate.name_group.set()
+        await call.message.answer(f'Введите название компании',
+                                  reply_markup=BotTools.construction_inline_keyboard(
+                                      buttons=['На главную'],
+                                      call_back=['start'],
+                                      message=call.message)
+                                  )
+        await call.answer()
+        await call.message.delete()
+    else:
+        await call.message.answer(f'Перед этим вам необходимо задать в настройках ваше имя пользователя\n\n'
+                                  f'Для этого вам необходимо пройти следующий путь:\n'
+                                  f'Настройки->Аккаунт->Имя пользователя',
+                                  reply_markup=BotTools.construction_inline_keyboard(
+                                      buttons=['На главную'],
+                                      call_back=['start'],
+                                      message=call.message))
+
+
+async def create_group_password(message: types.Message, state: FSMContext):
     async with state.proxy() as data:
-        data['group_id'] = message.text
-    await states_connect_to.StepsConnectTo.next()
-    await message.answer(f'Введите пароль от комнаты\n'
-                         f'(попросите вашего Мастера сообщить его вам, если этого не произошло ранее)',
+        data['name_group'] = message.text
+    await states_create_group.StepsCreate.next()
+    await message.answer(f'Введите пароль от комнаты\n',
                          reply_markup=BotTools.construction_inline_keyboard(
                              buttons=['На главную'],
                              call_back=['start'],
@@ -50,32 +53,57 @@ async def linc_group_password(message: types.Message, state: FSMContext):
                          )
 
 
-async def linc_group_check(message: types.Message, state: FSMContext):
-    from main import bot
+async def create_group_repeat_password(message: types.Message, state: FSMContext):
     async with state.proxy() as data:
         data['password'] = message.text
-    group = data['group_id']
-    password = data['password']
-    if con.work_with_MySQL(f'SELECT id FROM game_stories '
-                           f'WHERE id_group = {group} AND '
-                           f'password = {password}'):
-        expire_date = datetime.now() + timedelta(days=1)
-        link = await bot.create_chat_invite_link(group, expire_date.timestamp, 1)
-        # print(link)
+    await states_create_group.StepsCreate.next()
+    await message.answer(f'Повторите пароль от комнаты\n',
+                         reply_markup=BotTools.construction_inline_keyboard(
+                             buttons=['На главную'],
+                             call_back=['start'],
+                             message=message)
+                         )
 
-        await message.answer(f'Приятной игры!\n'
-                             f'Ваша ссылка для подключения к группе - {link.invite_link}',
-                             reply_markup=BotTools.construction_inline_keyboard(
-                                 buttons=['На главную'],
-                                 call_back=['start'],
-                                 message=message)
-                             )
+
+async def create_group_check(message: types.Message, state: FSMContext):
+    from main import bot
+    async with state.proxy() as data:
+        data['repeat_password'] = message.text
+    name = data['name_group']
+    password = data['password']
+    if password == data['repeat_password']:
+        bot_name = env.read_json_data('BOT_NAME')
+        username = message.from_user.username
+        await message.answer('Подождите, группа создаётся...')
+        try:
+            id_group = await pyro.supergroup_create(title=name, bot_name=bot_name, user_name=f'@{username}')
+            con.work_with_MySQL(f'INSERT INTO game_stories (GM_id, name_story, id_group, password) '
+                                f'VALUES ('
+                                f'(SELECT id FROM users WHERE user_id = "{message.from_user.id}"),'
+                                f'"{name}",'
+                                f'"{id_group}",'
+                                f'"{password}")')
+            await message.answer('Готово!',
+                                 reply_markup=BotTools.construction_inline_keyboard(
+                                     buttons=['На главную'],
+                                     call_back=['start'],
+                                     message=message))
+            await bot.send_message(id_group, f'ID компании: {id_group}\n'
+                                             f'Пароль: {password}\n')
+        except Exception as err:
+            await message.answer(f'Произошла ошибка!\n'
+                                 f'Ошибка - {err}\n\n'
+                                 f'Пожалуйста сообщете об этом администратору',
+                                 reply_markup=BotTools.construction_inline_keyboard(
+                                    buttons=['На главную'],
+                                    call_back=['start'],
+                                    message=message))
     else:
-        await message.answer('Неверный id или пароль группы!',
+        await message.answer('Пароли не совпадают!',
                              reply_markup=BotTools.construction_inline_keyboard(
                                  buttons=[['Попробовать сново'],
                                           ['На главную']],
-                                 call_back=[['connect_to'],
+                                 call_back=[['create_new_game'],
                                             ['start']],
                                  message=message))
-        await state.finish()
+    await state.finish()
